Test suite for all persisting services, to make sure they are executed in the right order (and this way, connecting to the database only needs to happen once)
